+++ src/cli/index.ts
@@ -5,6 +5,7 @@ import { run as runPreview } from './preview'
 import { runSSG, runSSR } from '../runner'
 import { rewritePackageJson } from '../init'
 import { toRootAbsolute } from '../utils'
+import { resolveConfig } from 'vite'
 
 yargs(hideBin(process.argv))
   .command(
@@ -38,14 +39,12 @@ yargs(hideBin(process.argv))
         describe: 'server side outputs directory',
         default: undefined
       }),
-    ({ outDirClient, outDirServer }) =>
-      runSSR({
-        outDirClient,
-        outDirServer,
-        build: {
-          ssrManifest: true
-        }
-      })
+    async () => {
+      const config = await resolveConfig({}, 'build')
+      config.build.ssrManifest = true
+
+      return runSSR(config)
+    }
   )
   .command(
     'generate',
@@ -58,11 +57,10 @@ yargs(hideBin(process.argv))
         describe: 'server side outputs directory',
         default: undefined
       }),
-    ({ outDirClient, outDirServer }) =>
-      runSSG({
-        outDirClient,
-        outDirServer
-      })
+    async () => {
+      const config = await resolveConfig({}, 'build')
+      return runSSG(config)
+    }
   )
   .command(
     'preview',