+++ src/constants.ts
@@ -1,6 +1,13 @@
 export const EXTENSIONS = ['jsx', 'tsx', 'ts', 'js', 'vue'] as const
 export const PIPED_EXTENSIONS = EXTENSIONS.join('|')
 export const NAME = 'vite:ssr-ssg'
+export const FRAMEWORKS = [
+  'react',
+  'preact',
+  'vue',
+  'svelte',
+  'vanilla'
+] as const
 
 export const CYAN = '\u001b[36m'
 export const GREEN = '\u001b[32m'