+++ src/react/routes.tsx
@@ -1,9 +1,33 @@
 import { ComponentType } from 'react'
+const bracketRegex = /\[.+\]\..+$/
 
-type GetRoutes = {
+type Routes = {
   name: string
   path: string
   Component: ComponentType
+}
+
+const path2RouteObject = (
+  path: string
+): {
+  name: string
+  path: string
+} => {
+  const fileName = path.split('/').slice(-1)[0]
+  const isDynamicPath = bracketRegex.test(fileName)
+  const name = isDynamicPath
+    ? fileName.split('.')[0].replace(/\[/, '').replace(/\]/, '')
+    : fileName.split('.')[0]
+  const _path = isDynamicPath
+    ? `/:${name}`
+    : name.toLowerCase() === 'index'
+    ? '/'
+    : `/${name?.toLowerCase()}`
+
+  return {
+    name,
+    path: _path
+  }
 }
 
 const getRoutes = (
@@ -13,16 +37,15 @@ const getRoutes = (
       [key: string]: ComponentType
     }
   >,
-  key = 'default'
-): GetRoutes[] =>
+  module = 'default'
+): Routes[] =>
   Object.entries(pages).map(([path, fn]) => {
-    const fileName = path.split('/').slice(-1)[0]
-    const name = fileName.split('.')[0]
+    const { name, path: _path } = path2RouteObject(path)
 
     return {
       name,
-      path: name === 'Index' ? '/' : `/${name?.toLowerCase()}`,
-      Component: fn[key]
+      path: _path,
+      Component: fn[module]
     }
   })