+++ src/vue3/template.ts
@@ -41,11 +41,11 @@ import {
   createRouter,
   createWebHistory${isTS ? ',\n  Router' : ''}
 } from 'vue-router'
-import { getRoutes } from 'vite-plugin-ssr-ssg/vue3'
+import { getRoutes } from 'vite-plugin-ssr-ssg'
 import { createHead${isTS ? ', HeadClient' : ''} } from '@vueuse/head'
 
 const pages = import.meta.globEager('./pages/*.vue')
-const routes = getRoutes(pages)
+const routes = getRoutes${isTS ? "<'vue'>" : ''}(pages)
 
 export const createApp = ()${isTS ? type : ''} => {
   const app = createSSRApp(App)